[A16] camera.bst: fix 64-bit pointer truncation in MessageQueue arg0
CMessage::arg0 was int32_t, used to pass ConfigureParam* pointer via
VideoStream::postMessage(new CMessage(MSG_CONFIG, (uintptr_t)params)).
On 64-bit, the pointer is truncated to 32 bits when stored in arg0,
causing a crash in VideoStream::handleConfigureLocked when the invalid
pointer is dereferenced.

Fix: change arg0 from int32_t to uintptr_t (port of BAM-544 fix).
This only manifests on 64-bit-only builds where the camera provider
runs as 64-bit; on 32/64 the 32-bit provider was unaffected.

+++ hardware/bst/camera/3.0/MessageQueue.h
@@ -62,10 +62,10 @@ public:
 class CMessage : public LightRefBase<CMessage>{
 public:
     int32_t what;
-    int32_t arg0;
+    uintptr_t arg0;
 
     CMessage(int32_t what,
-             int32_t arg0 = 0)
+             uintptr_t arg0 = 0)
         : what(what), arg0(arg0) {}
 
     virtual ~CMessage() {}
